backend/requirements.txt
fastapi
uvicorn
langchain-community
langchain-openai
langgraph
pydantic
python-dotenv
semantic-chunkers
semantic-router
tavily-python
python-multipart
httpx
sqlalchemy
asyncpg
aiosqlite
alembic
python-jose[cryptography]
